% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcrossPedigree.R
\name{backcrossPedigree}
\alias{backcrossPedigree}
\title{Generate a backcross pedigree which starts from inbred founders}
\usage{
backcrossPedigree(populationSize)
}
\arguments{
\item{populationSize}{The size of the generated population.}
}
\value{
An object of class \code{detailedPedigree} representing the experimental design, suitable for simulation using simulateMPCross.
}
\description{
Generate a backcross pedigree which starts from inbred founders
}
\details{
Generate a backcross pedigree which starts from inbred founders
}
\examples{
pedigree <- backcrossPedigree(1000)
#This pedigree is automatically marked as involving finite generations of selfing. 
selfing(pedigree)
}
