% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callFromMap.R
\name{callFromMap}
\alias{callFromMap}
\title{Call markers based on an existing map}
\usage{
callFromMap(rawData, thresholdChromosomes = 100,
  thresholdAlleleClusters = c(1e-10, 1e-20, 1e-30, 1e-40),
  maxChromosomes = 2, existingImputations, tDistributionPValue = 0.6,
  useOnlyExtraImputationPoints = TRUE)
}
\arguments{
\item{rawData}{Raw data for a genetic marker.}

\item{thresholdChromosomes}{The test-statistic threshold for declaring a marker to be polymorphic on a chromosome.}

\item{thresholdAlleleClusters}{The p-value threshold for declaring two underlying founder alleles to have different marker alleles. Multiple possible values should be input.}

\item{maxChromosomes}{The maximum number of chromosomes that a marker can be polymorphic on}

\item{existingImputations}{An object of class mpcrossMapped from the mpMap2 package, containing data about imputed underlying genotypes.}

\item{tDistributionPValue}{Paramater controlling the size of each detected cluster, ranging from 0 to 1. Small values result in small clusters, and large values result in large clusters.}
}
\description{
This function uses an existing genetic map to call genetic markers, including markers polymorphic on multiple chromosomes.
}
\details{
This function uses an existing genetic map to call a genetic marker. There are a number of advantages to this approach
\describe{
	\item{1.}{It can correctly call markers which are polymorphic on multiple chromosomes, therefore converting one marker into two.}
\item{2.}{It avoids incorrectly calling markers polymorphic on multiple chromosomes. Incorrect calling can lead to supurious genetic interactions.}
	\item{3.}{It can call markers that initially appear to be monomorphic in the population. }
	\item{4.}{It can call additional marker alleles for markers that would otherwise be ignored. }
}

Once a genetic map has been constructed, it should be used to impute underlying founder genotypes at an equally spaced grid of points using function \code{\link[mpMap2]{imputeFounders}}. 
The steps in the algorithm are as follows:
\describe{
	\item{1.}{Determine which chromosomes the marker is associated to, and where on those chromosomes. This is determined using function \code{\link[mpMap2]{addExtraMarkerFromRawCall}}, which is itself based on a manova model. 
  The marker is assumed associated to chromosomes for which the test statistic is greater than \code{thresholdChromosomes}. 
	  An appropriate value for \code{thresholdChromosomes} can be determined by looking at the results of \code{\link[mpMap2]{addExtraMarkerFromRawCall}}, for a number of different markers.}
	\item{2.}{Determine the distribution of marker alleles, at all the associated genetic locations. 
   This is done by taking the founders to be the vertices of a graph, and connecting founders which seem to part of the same marker allele. 
	  The resulting graph should be a union of disjoint complete graphs (cliques). }
	\item{3.}{We now have a preliminary assignment of marker alleles to lines, where the assignment may be of 1, 2, 3 or more \emph{different} marker alleles, depending on how many chromosomes the marker is associated with.
	  For example, if the marker is associated with two chromosomes, then there will be two marker alleles for each line. 
	  For each unique combination of marker alleles, we take the lines which have that assignment of marker alleles, and fit a skew-t distribution. }
	\item{4.}{For each fitted distribution, determine a confidence region using p-value \code{tDistributionPValue}. }
	\item{5.}{Use these confidence regions to construct marker calls at each associated location.}
}
}
