% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeMarkerPosition.R
\name{changeMarkerPosition}
\alias{changeMarkerPosition}
\title{Change the position of a single marker}
\usage{
changeMarkerPosition(mpcrossMapped, marker, newChromosome, newPosition)
}
\arguments{
\item{mpcrossMapped}{The object of class \code{mpcrossMapped}, containing a marker to be modified}

\item{marker}{The name of the marker to change}

\item{newChromosome}{The new chromosome for the specified marker}

\item{newPosition}{The new position for the specified marker in cM, on the new chromosome}
}
\value{
A copy of the input object, with the specified marker shifted to the new position and chromosome.
}
\description{
Change the position of a single marker
}
\details{
For an object of class \code{mpcrossMapped}, change the position of a single marker
}
