% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAllEpistaticChiSquared.R
\name{computeAllEpistaticChiSquared}
\alias{computeAllEpistaticChiSquared}
\title{Compute chi-squared test statistics for independence}
\usage{
computeAllEpistaticChiSquared(mpcrossMapped, verbose = TRUE)
}
\arguments{
\item{mpcrossMapped}{An object of class \code{mpcrossMapped} with IBD probability data.}

\item{verbose}{If this is \code{TRUE} a progress bar is generated}
}
\value{
A square matrix with rows and columns corresponding to genetic locations, and values corresponding to test statistics.
}
\description{
Compute chi-squared test statistics for independence
}
\details{
This function computes what are (approximately) chi-squared test statistics for independence of the genotypes at different points on the genome. This computation is done using the IBD probability data. Significant non-independence between IBD probabilities at distant points on the same chromosome, or points on different chromosomes, can indicate non-standard genetic inheritance or selective pressure.
}
