% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eightWayPedigreeImproperFunnels.R
\name{eightParentPedigreeImproperFunnels}
\alias{eightParentPedigreeImproperFunnels}
\title{Generate an eight-parent pedigree}
\usage{
eightParentPedigreeImproperFunnels(initialPopulationSize, selfingGenerations,
  nSeeds)
}
\arguments{
\item{initialPopulationSize}{The number of F1 lines generated}

\item{selfingGenerations}{The number of selfing generations at the end of the pedigree}

\item{nSeeds}{The number of progeny taken from each intercrossing line, or from each F1 if no intercrossing is specified. These lines are then selfed according to selfingGenerations}
}
\description{
Generate a eight-parent pedigree starting from inbred founders, where the founders in the funnels are not necessarily distinct.
}
\examples{
pedigree <- eightParentPedigreeImproperFunnels(initialPopulationSize = 10, 
	selfingGenerations = 0, nSeeds = 1)
#convert the pedigree to a graph
pedigreeAsGraph <- pedigreeToGraph(pedigree)
#Plot it
\dontrun{plot(pedigreeAsGraph)}
#Write it to a file in DOT format
}
\seealso{
\code{\link{eightParentPedigreeSingleFunnel}}, \code{\link{fourParentPedigreeSingleFunnel}}, \code{\link{fourParentPedigreeRandomFunnels}}, \code{\link{twoParentPedigree}}
}

