% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eightWayPedigreeRandomFunnels.R
\name{eightParentPedigreeRandomFunnels}
\alias{eightParentPedigreeRandomFunnels}
\title{Generate an eight-parent pedigree}
\usage{
eightParentPedigreeRandomFunnels(initialPopulationSize, selfingGenerations,
  nSeeds = 1L, intercrossingGenerations)
}
\arguments{
\item{initialPopulationSize}{The number of F1 lines generated}

\item{selfingGenerations}{The number of selfing generations at the end of the pedigree}

\item{nSeeds}{The number of progeny taken from each intercrossing line, or from each F1 if no intercrossing is specified. These lines are then selfed according to selfingGenerations}

\item{intercrossingGenerations}{The number of generations of random mating performed from the F1 generation. Population size is maintained at that specified by initialPopulationSize}
}
\description{
Generate a eight-parent pedigree starting from inbred founders, using a random funnel
}
\examples{
pedigree <- eightParentPedigreeRandomFunnels(initialPopulationSize = 10, 
	selfingGenerations = 0, nSeeds = 1, intercrossingGenerations = 10)
#convert the pedigree to a graph
pedigreeAsGraph <- pedigreeToGraph(pedigree)
#Plot it
\dontrun{plot(pedigreeAsGraph)}
#Write it to a file in DOT format
}
\seealso{
\code{\link{eightParentPedigreeSingleFunnel}}, \code{\link{fourParentPedigreeSingleFunnel}}, \code{\link{fourParentPedigreeRandomFunnels}}, \code{\link{twoParentPedigree}}
}

