% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{eightParentSubsetMap}
\alias{eightParentSubsetMap}
\title{Genetic map and genetic data from an 8-parent MAGIC population.}
\description{
Genetic map and genetic data from an 8-parent MAGIC population.
}
\author{
Alex Whan, Matthew Morell, Rohan Shah, Colin Cavanagh
This dataset contains the genetic map, genetic data, and imputed IBD genotypes for parts of chromosomes 1A, 1B and 1D, from an 8-way MAGIC population of 4229 lines.
}
