% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{eightWayExampleData}
\alias{eightWayExampleData}
\title{Example data from an 8-parent MAGIC population.}
\description{
Example data from an 8-parent MAGIC population.
}
\author{
Alex Whan, Matthew Morell, Rohan Shah, Colin Cavanagh
This is a sample of 200 genetic markers from an 8-way MAGIC population of 4229 lines.
}
