% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMap.R
\name{estimateMap}
\alias{estimateMap}
\title{Computes map distances}
\usage{
estimateMap(mpcrossLG, mapFunction = rfToHaldane, maxOffset = 1,
  maxMarkers = 2000, verbose = FALSE)
}
\arguments{
\item{maxMarkers}{The (approximate) number of markers for which distances are estimated simultaneously. 
Use the non-linear least squares function to estimate a map}
}
\description{
Computes map distances
}

