% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRF.R
\name{estimateRF}
\alias{estimateRF}
\title{Estimate recombination fractions}
\usage{
estimateRF(object, recombValues, lineWeights, gbLimit = -1, keepLod = FALSE,
  keepLkhd = FALSE, verbose = FALSE, markerRows = 1:nMarkers(object),
  markerColumns = 1:nMarkers(object))
}
\arguments{
\item{object}{The input mpcross object}

\item{recombValues}{a vector of test values to use for the numeric maximum likelihood step. Must contain 0 and 0.5, and must have less than 255 values in total. The default value is \code{c(0:20/200, 11:50/100)}.}

\item{lineWeights}{Values to use to correct for segregation distortion. This parameter should in general be left unspecified.}

\item{gbLimit}{The maximum amount of working memory this estimation step should be allowed to use at any one time, in gigabytes. Smaller values may increase the computation time. A value of -1 indicates no limit.}

\item{keepLod}{Set to \code{TRUE} to compute the likelihood ratio score statistics for testing whether the estimate is different from 0.5. Due to memory constraints this should generally be left as \code{FALSE}.}

\item{keepLkhd}{Set to \code{TRUE} to compute the maximum value of the likelihood. Due to memory constraints this should generally be left as \code{FALSE}.}

\item{verbose}{Output diagnostic information, such as the amount of memory required, and the progress of the computation}

\item{markerRows}{Used to estimate only a subset of the full matrix of pairwise recombination fractions.}

\item{markerColumns}{Used to estimate only a subset of the full matrix of pairwise recombination fractions.}

\item{lineWeights}{Used to correct for the effects of segregation distortion.}
}
\value{
An object of class \code{mpcrossRF}, which contains the original genetic data, and also estimated recombination fraction data.
}
\description{
This function estimates the recombination fractions between all pairs of markers in the input object. The recombination fractions are estimated using numerical maximum likelihood, and a grid search. Because every estimate will be one of the input test values, the estimates can be stored efficiently with a single byte per estimate.
}
\examples{
map <- qtl::sim.map(len = 100, n.mar = 11, include.x=FALSE)
f2Pedigree <- f2Pedigree(1000)
cross <- simulateMPCross(map = map, pedigree = f2Pedigree, mapFunction = haldane, seed = 1)
rf <- estimateRF(cross)
#Print the estimated recombination fraction values
rf@rf@theta[1:11, 1:11]
}

