% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{existingLocalisationStatistics}
\alias{existingLocalisationStatistics}
\title{Localisation statistics for example of callFromMap}
\description{
This dataset contains the localisation statistics for the example for running \code{callFromMap}. This makes the example fast enough to pass the CRAN check.
}
