% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Expand markers contained within object}
\usage{
expand(mpcross, newMarkers)
}
\arguments{
\item{mpcross}{The input \code{mpcross} object}

\item{newMarkers}{The names of the new markers to add}
}
\value{
An object of class \code{mpcross} with a larger set of markers.
}
\description{
Expand set of markers within object, adding extra markers with missing observations as necessary.
}
\details{
This function expands the set of markres within an \code{mpcross} object. The new set of marker names must contain all the existing marker names, with any desired extra marker names. Any added markers will have all observations marked as missing. Any existing non-genetic information (genetic map, assignment of linkage groups, IBD genotypes, IBD probabilitieS) will be removed.
}
