% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportMapToPretzl.R
\name{exportMapToPretzl}
\alias{exportMapToPretzl}
\title{Export genetic map to Pretzl}
\usage{
exportMapToPretzl(inputObject, name, separateChromosomes = FALSE)
}
\arguments{
\item{inputObject}{The object of class \code{mpcrossMapped} containing the genetic map}

\item{name}{If a single JSON object is being exported, the name of the exported map.}

\item{separateChromosomes}{If \code{TRUE}, separate exports will be generated for each chromosome. The name associated with each chromosome map will contain the chromosome name as a suffix.}
}
\value{
A list containing JSON, suitable for import into Pretzl.
}
\description{
Export genetic map to Pretzl
}
\details{
Convert the genetic map from an object of class \code{mpcrossMapped} to the JSON format used by Pretzl. Pretzl is a web app for visualising and comparing genetic maps.
}
