% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraImputationPoints.R
\name{extraImputationPoints}
\alias{extraImputationPoints}
\title{Get out non-marker positions used for IBD genotype imputation}
\usage{
extraImputationPoints(mpcrossMapped)
}
\arguments{
\item{mpcrossMapped}{The object from which to get the non-marker positions}
}
\value{
A vector of genetic position names.
}
\description{
Get out non-marker positions used for IBD genotype imputation
}
\details{
Extract non-marker positions used for IBD genotype imputation
}
