% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2Pedigree.R
\name{f2Pedigree}
\alias{f2Pedigree}
\title{Generate an F2 pedigree which starts from inbred founders}
\usage{
f2Pedigree(populationSize)
}
\arguments{
\item{populationSize}{The size of the generated population.}
}
\value{
An object of class \code{detailedPedigree} representing the experimental design, suitable for simulation using simulateMPCross.
}
\description{
Generate an F2 pedigree which starts from inbred founders
}
\examples{
pedigree <- f2Pedigree(1000)
#This pedigree is automatically marked as involving finite generations of selfing. 
selfing(pedigree)
}
