% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finals.R
\name{finalNames}
\alias{finalNames}
\alias{finalNames,mpcross-method}
\alias{finalNames,geneticData-method}
\title{Names of genetic lines

Return the names of the genetic lines

If the \code{mpcross} object contains a single experiment a vector of names of genetic lines is returned. The names of the founding lines for the population are excluded. If an \code{mpcross} object contains multiple experiments a list of vectors of names is returned.}
\usage{
finalNames(object)

\S4method{finalNames}{mpcross}(object)

\S4method{finalNames}{geneticData}(object)
}
\arguments{
\item{object}{The \code{mpcross} object from which to extract the names of the genetic lines}
}
\value{
The names of the genetic lines in the final population.
}
\description{
Names of genetic lines

Return the names of the genetic lines

If the \code{mpcross} object contains a single experiment a vector of names of genetic lines is returned. The names of the founding lines for the population are excluded. If an \code{mpcross} object contains multiple experiments a list of vectors of names is returned.
}
