% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationGenerics.R
\name{flatImputationMapNames}
\alias{flatImputationMapNames}
\alias{flatImputationMapNames,imputed-method}
\alias{flatImputationMapNames,geneticData-method}
\alias{flatImputationMapNames,mpcrossMapped-method}
\title{Get names of positions for IBD genotype imputation}
\usage{
flatImputationMapNames(object, ...)

\S4method{flatImputationMapNames}{imputed}(object, ...)

\S4method{flatImputationMapNames}{geneticData}(object, ...)

\S4method{flatImputationMapNames}{mpcrossMapped}(object, ...)
}
\arguments{
\item{object}{The object from which to get the names of positions}

\item{...}{Extra parameters, currently only \code{"experiment"} is supported.}
}
\value{
The names of all positions at which IBD genotype imputation has already been performed.
}
\description{
Get the names of all positions at which IBD genotype imputation has already been performed
}
\details{
Get the names of all positions at which IBD genotype imputation has already been performed
}
