% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/founders.R
\name{founderNames}
\alias{founderNames}
\title{Names of founding genetic lines

Return the names of the founding genetic lines

If the \code{mpcross} object contains a single experiment a vector of names of genetic lines is returned. If an \code{mpcross} object contains multiple experiments a list of vectors of names is returned.}
\usage{
founderNames(object)
}
\description{
Names of founding genetic lines

Return the names of the founding genetic lines

If the \code{mpcross} object contains a single experiment a vector of names of genetic lines is returned. If an \code{mpcross} object contains multiple experiments a list of vectors of names is returned.
}
