% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/founders.R
\name{founders}
\alias{founders}
\title{Genetic data for founding lines

Return the genetic data matrix for the founding lines

If the \code{mpcross} object contains a single experiment a matrix is returned, with rows corresponding to founding lines and columns corresponding to markers. If an \code{mpcross} object contains multiple experiments a list of such matrices is returned, one for each experiment.}
\usage{
founders(object)
}
\description{
Genetic data for founding lines

Return the genetic data matrix for the founding lines

If the \code{mpcross} object contains a single experiment a matrix is returned, with rows corresponding to founding lines and columns corresponding to markers. If an \code{mpcross} object contains multiple experiments a list of such matrices is returned, one for each experiment.
}
