% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcross.R
\name{fromMpMap}
\alias{fromMpMap}
\title{Convert from mpMap format to mpMap2 format}
\usage{
fromMpMap(mpcross, selfing = "infinite", fixCodingErrors = FALSE)
}
\arguments{
\item{mpcross}{Object to convert}

\item{selfing}{Number of generations of selfing to put in the pedigree, for the converted object. Must be \code{"finite"} or \code{"infinite"}.}

\item{fixCodingErrors}{Should we attempt to fix coding errors, by replacing invalid values by NA? Defaults to FALSE.}
}
\value{
An object of class \code{mpcross} or \code{mpcrossMapped}, depending on the data contained in the input object.
}
\description{
Convert an object from mpMap format into mpMap2 format
}
\details{
Convert an object from mpMap format (the predecessor to mpMap2) into mpMap2 format. It is unlikely that this function will ever need to be used.
}
