% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPositions.R
\name{getPositions}
\alias{getPositions}
\title{Get positions of genetic markers}
\usage{
getPositions(mpcrossMapped, markers)
}
\arguments{
\item{mpcrossMapped}{The \code{mpcross} object containing the map of interest}

\item{markers}{The markers for which to get the positions}
}
\description{
Get positions of genetic markers, on their respective chromosomes
}
\details{
Get positions of genetic markers in cM, on their respective chromosomes
}
