% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetData.R
\name{hetData}
\alias{hetData}
\alias{hetData,mpcross-method}
\alias{hetData,geneticData-method}
\title{Get the encoding of marker heterozygotes}
\usage{
hetData(object, marker)

\S4method{hetData}{mpcross}(object, marker)

\S4method{hetData}{geneticData}(object, marker)
}
\arguments{
\item{object}{The object from which to extract the encoding data}

\item{marker}{The marker of interest. If this is missing, heterozygote encoding data is returned for all markers.}
}
\value{
Heterozygote encoding data, for either a specific marker or all markers.
}
\description{
Get the encoding of marker heterozygotes
}
\details{
Get the encoding of markers heterozygotes, either for all markers, or a specific marker.
}
