% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcross-class.R
\name{hetsForSNPMarkers}
\alias{hetsForSNPMarkers}
\title{Create heterozygote encodings for SNP markers}
\usage{
hetsForSNPMarkers(founders, finals, pedigree)
}
\arguments{
\item{founders}{Genetic data for the founding lines of the population}

\item{finals}{Genetic data for the final genotyped lines of the population}

\item{pedigree}{Pedigree for the population. Unused by this particular function.}
}
\value{
An object of class \code{hetData}, which contains encodings for the marker heterozygotes and the (unique) marker heterozygote
}
\description{
Create encoding which assumes that the single non-homozygote value for a SNP marker is the heterozygote
}
\details{
This function takes in genotype data for the founding lines and the final poulation. It returns an encoding for hetorozygotes for all markers, where multiallelic markers are assumed to have no heterozygotes. For biallelic markers with three observed alleles in the final population, the extra allele is assumed to be the heterozygote.
}
