% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canSkipValidity.R, R/geneticData-class.R
\name{initialize,canSkipValidity-method}
\alias{initialize,canSkipValidity-method}
\alias{initialize,geneticDataList-method}
\title{Initialize method which can skip the validity check}
\usage{
\S4method{initialize}{canSkipValidity}(.Object, ...)

\S4method{initialize}{geneticDataList}(.Object, ...)
}
\arguments{
\item{.Object}{the object to initialize}

\item{...}{Other arguments. Only \code{skipValidity} is used.}
}
\description{
This is an initialization method with an optional \code{skipValidity} argument. If this argument is set to \code{TRUE}, the validity check is skipped. This is used by some internal functions within the package, as the validity check can be slow, and internal code is (presumably) guaranteed to produce valid objects.
}
\details{
Initialize method which can skip the validity check
}
