% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawSymmetricMatrix.R, R/assignFounderPattern.R,
%   R/assignFounderPatternPrototype.R, R/biparentalDominant.R,
%   R/fixedNumberOfFounderAlleles.R, R/multiparentSNP.R,
%   R/multiparentSNPPrototype.R, R/removeHets.R, R/simulatePhenotypes.R
\name{[,rawSymmetricMatrix,index,index,logical-method}
\alias{[,rawSymmetricMatrix,index,index,logical-method}
\alias{[,rawSymmetricMatrix,index,index,missing-method}
\alias{[,rawSymmetricMatrix,missing,missing,missing-method}
\alias{[,rawSymmetricMatrix,matrix,missing,missing-method}
\alias{+,geneticData,assignFounderPattern-method}
\alias{+,mpcross,assignFounderPattern-method}
\alias{+,mpcrossMapped,assignFounderPattern-method}
\alias{+,geneticData,assignFounderPatternPrototype-method}
\alias{+,mpcross,assignFounderPatternPrototype-method}
\alias{+,mpcrossMapped,assignFounderPatternPrototype-method}
\alias{+,geneticData,biparentalDominant-method}
\alias{+,mpcross,biparentalDominant-method}
\alias{+,geneticData,fixedNumberOfFounderAlleles-method}
\alias{+,mpcross,fixedNumberOfFounderAlleles-method}
\alias{+,geneticData,multiparentSNP-method}
\alias{+,mpcross,multiparentSNP-method}
\alias{+,geneticData,multiparentSNPPrototype-method}
\alias{+,mpcross,multiparentSNPPrototype-method}
\alias{+,mpcross,removeHets-method}
\alias{+,geneticData,normalPhenotype-method}
\alias{+,mpcross,normalPhenotype-method}
\title{Internal operators for mpMap2}
\usage{
\S4method{[}{rawSymmetricMatrix,index,index,logical}(x, i, j, ..., drop = TRUE)

\S4method{[}{rawSymmetricMatrix,index,index,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{rawSymmetricMatrix,missing,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{rawSymmetricMatrix,matrix,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{+}{geneticData,assignFounderPattern}(e1, e2)

\S4method{+}{mpcross,assignFounderPattern}(e1, e2)

\S4method{+}{mpcrossMapped,assignFounderPattern}(e1, e2)

\S4method{+}{geneticData,assignFounderPatternPrototype}(e1, e2)

\S4method{+}{mpcross,assignFounderPatternPrototype}(e1, e2)

\S4method{+}{mpcrossMapped,assignFounderPatternPrototype}(e1, e2)

\S4method{+}{geneticData,biparentalDominant}(e1, e2)

\S4method{+}{mpcross,biparentalDominant}(e1, e2)

\S4method{+}{geneticData,fixedNumberOfFounderAlleles}(e1, e2)

\S4method{+}{mpcross,fixedNumberOfFounderAlleles}(e1, e2)

\S4method{+}{geneticData,multiparentSNP}(e1, e2)

\S4method{+}{mpcross,multiparentSNP}(e1, e2)

\S4method{+}{geneticData,multiparentSNPPrototype}(e1, e2)

\S4method{+}{mpcross,multiparentSNPPrototype}(e1, e2)

\S4method{+}{mpcross,removeHets}(e1, e2)

\S4method{+}{geneticData,normalPhenotype}(e1, e2)

\S4method{+}{mpcross,normalPhenotype}(e1, e2)
}
\arguments{
\item{x}{object from which to extract element(s)}

\item{i, j}{indices specifying elements to extract or replace}

\item{...}{Currently unused}

\item{drop}{If TRUE the result is coerced to the lowest possible dimension}

\item{e1}{Object one}

\item{e2}{Object two}
}
\value{
Various. Not for external use.

Modified version of the input object. The class of the output depends on the class of the input.
}
\description{
Internal operators, used to modify mpcross objects.
}
\details{
These operators are used to combine objects in order to apply a function. For example, \code{e1 + biparentalDominant()} returns the biparental design \code{e1}, with all markers converted to dominant markers. Consult the documentation on the individual functions, rather than this list of operators.
}
