% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineNames.R
\name{lineNames}
\alias{lineNames}
\title{Get or set the genetic line names}
\usage{
lineNames(object)
}
\arguments{
\item{object}{The object from which to extract the line names}
}
\value{
Vector of genetic line names
}
\description{
Get or set the genetic line names associated with a pedigree or \code{mpcross} object.
}
\details{
These functions get or set the names of the genetic lines associated with a pedigree or \code{mpcross} object.
}
