% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineNames.R
\docType{methods}
\name{lineNames,mpcross-method}
\alias{lineNames,mpcross-method}
\alias{lineNames,geneticData-method}
\title{Get the genetic line names}
\usage{
\S4method{lineNames}{mpcross}(object)

\S4method{lineNames}{geneticData}(object)
}
\arguments{
\item{object}{The object from which to extract the line names}
}
\description{
Get the genetic line names of a population
}
\details{
These functions get the names of the genetic lines associated with an \code{mpcross} object.
}
