% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineNames.R
\name{lineNames<-}
\alias{lineNames<-}
\alias{lineNames,pedigree-method}
\alias{lineNames<-,detailedPedigree-method}
\alias{lineNames<-,pedigree-method}
\title{Get or set the genetic line names of a pedigree}
\usage{
lineNames(object) <- value

\S4method{lineNames}{pedigree}(object)

\S4method{lineNames}{detailedPedigree}(object) <- value

\S4method{lineNames}{pedigree}(object) <- value
}
\arguments{
\item{object}{The object for which to get or set the line names}

\item{value}{The vector of new genetic line names}
}
\value{
None
}
\description{
Get or set the genetic line names of a pedigree
}
\details{
These functions get or set the names of the genetic lines associated with a pedigree.
}
