% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{linesByNames}
\alias{linesByNames}
\title{Extract pedigree by names}
\usage{
linesByNames(pedigree, names)
}
\arguments{
\item{pedigree}{An object of class \code{pedigree}}

\item{names}{The names of the lines for which to extract the pedigree}
}
\value{
A matrix giving the genetic lines and their parents, by line name.
}
\description{
Extract part of pedigree in human-readable format
}
\details{
Pedigrees in mpMap2 are stored using indices for maternal and paternal lines, which is not a human-readable format. This function takes in a pedigree, and returns a human-readable subset.
}
