% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFunctions.R
\name{mapFunctions}
\alias{haldane}
\alias{haldaneToRf}
\alias{kosambi}
\alias{kosambiToRf}
\alias{mapFunctions}
\alias{rfToHaldane}
\alias{rfToKosambi}
\title{Map functions}
\usage{
haldaneToRf(x)

haldane(x)

rfToHaldane(r)

rfToKosambi(r)

kosambiToRf(x)

kosambi(x)
}
\arguments{
\item{x}{centiMorgan distance}

\item{r}{recombination fraction}

\item{r}{recombination fraction}

\item{x}{centiMorgan distance}
}
\section{Functions}{
\itemize{
\item \code{haldaneToRf}: Convert from Haldane distance to recombination fraction

\item \code{haldane}: Convert from Haldane distance to recombination fraction

\item \code{rfToHaldane}: Convert from recombination fraction to Haldane distance

\item \code{rfToKosambi}: Convert from recombination fraction to Kosambi distance

\item \code{kosambiToRf}: Convert from Kosambi distance to recombination fraction

\item \code{kosambi}: Convert from recombination fraction to Kosambi distance
}}

