% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markers}
\alias{markers}
\title{Genotyped markers

Return the names of the genotyped markers. 

If an \code{mpcross} object contains multiple experiments, all experiments are required to have the same markers. So a single vector of marker names is returned, in all cases.}
\usage{
markers(object)
}
\description{
Genotyped markers

Return the names of the genotyped markers. 

If an \code{mpcross} object contains multiple experiments, all experiments are required to have the same markers. So a single vector of marker names is returned, in all cases.
}
