% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcross-class.R
\docType{class}
\name{mpcross-class}
\alias{mpcross-class}
\alias{.mpcross}
\title{A collection of multi-parent populations without a genetic map}
\description{
A collection of multi-parent populations without a genetic map
}
\details{
An object of class \code{mpcross} contains data about one or more multi-parent populations, without a genetic map. As there is no genetic map, there is no information about IBD imputed genotypes or IBD genotype probabilities. There is also no information about estimated recombination fractions.
}
\section{Slots}{

\describe{
\item{\code{geneticData}}{A list of objects of class \code{geneticData}, each representing a population.}
}}

