% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcross-class.R
\docType{class}
\name{mpcrossMapped-class}
\alias{mpcrossMapped-class}
\alias{.mpcrossMapped}
\title{A collection of multi-parent populations with a genetic map}
\description{
A collection of multi-parent populations with a genetic map
}
\details{
An object of class \code{mpcrossMapped} contains genetic data for one or more populations, and a genetic map. It might also contain data about the recombination fractions between markers (or it might not).
}
\section{Slots}{

\describe{
\item{\code{map}}{The genetic map for all populations}

\item{\code{rf}}{The recombination fraction data (which might be NULL).}

\item{\code{geneticData}}{A list of objects of class \code{geneticData}, each representing a population.}
}}

