% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcross-class.R
\docType{class}
\name{mpcrossRF-class}
\alias{mpcrossRF-class}
\alias{.mpcrossRF}
\title{A collection of multi-parent populations with recombination fraction estimates}
\description{
A collection of multi-parent populations with recombination fraction estimates
}
\details{
An object of class \code{mpcrossRF} contains data about one or more multi-parent populations, without a genetic map, but with recombination fraction estimates. As there is no genetic map, there is no information about IBD imputed genotypes or IBD genotype probabilities.
}
\section{Slots}{

\describe{
\item{\code{geneticData}}{A list of objects of class \code{geneticData}, each representing a population.}

\item{\code{rf}}{Estimates of recombination fractions between every pair of genetic markers.}
}}

