% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiparentSNP.R
\name{multiparentSNP}
\alias{multiparentSNP}
\title{Convert all markers to SNP markers}
\usage{
multiparentSNP(keepHets)
}
\arguments{
\item{keepHets}{Should heterozygotes for the SNP marker be kept?}
}
\value{
An object of internal type \code{multiparentSNP}, which can be combined with an object of class \code{mpcross} using the addition operator.
}
\description{
Convert all markers in an object with fully informative markers to SNP markers
}
\details{
When initially generated, objects of class \code{mpcross} have markers that are fully informative - Every founder carries a different allele, and all marker heterozygotes are distinguishable. This function can be used to convert a simulated object to one with SNP markers. The resulting markers have two alleles, and the marker heterozygote might or might be observable.
}
