% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nFounders.R
\name{nFounders}
\alias{nFounders}
\title{Number of genotyped markers

Return the number of genotyped markers in an object.

If an \code{mpcross} object contains multiple experiments, one number is returned per experiment.}
\usage{
nFounders(object)
}
\description{
Number of genotyped markers

Return the number of genotyped markers in an object.

If an \code{mpcross} object contains multiple experiments, one number is returned per experiment.
}
