% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nLines.R, R/nMarkers.R
\docType{methods}
\name{nLines}
\alias{nLines}
\alias{nLines,mpcross-method}
\alias{nLines,geneticData-method}
\alias{nMarkers}
\title{Number of genotyped lines

Return the number of genotyped lines in an object.

This includes only the number of final lines genotyped in the population, and does not include the founding lines. If an \code{mpcross} object contains multiple experiments, one number is returned per experiment.}
\usage{
nLines(object)

\S4method{nLines}{mpcross}(object)

\S4method{nLines}{geneticData}(object)

nMarkers(object)
}
\description{
Number of genotyped lines

Return the number of genotyped lines in an object.

This includes only the number of final lines genotyped in the population, and does not include the founding lines. If an \code{mpcross} object contains multiple experiments, one number is returned per experiment.

Number of genotyped markers

Return the number of genotyped markers in an object.

If an \code{mpcross} object contains multiple experiments, all experiments are required to have the same markers. So only one number is returned, in all cases.
}
