% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nMarkers.R
\name{nMarkers}
\alias{nMarkers}
\alias{nMarkers,mpcross-method}
\alias{nMarkers,geneticData-method}
\title{Number of genotyped markers

Return the number of genotyped markers in an object.

If an \code{mpcross} object contains multiple experiments, all experiments are required to have the same markers. So only one number is returned, in all cases.}
\usage{
nMarkers(object)

\S4method{nMarkers}{mpcross}(object)

\S4method{nMarkers}{geneticData}(object)
}
\arguments{
\item{object}{The \code{mpcross} object from which to extract the marker names}
}
\value{
The number of markers in an object of class \code{mpcross}.
}
\description{
Number of genotyped markers

Return the number of genotyped markers in an object.

If an \code{mpcross} object contains multiple experiments, all experiments are required to have the same markers. So only one number is returned, in all cases.
}
