% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhenotypes.R
\name{normalPhenotype}
\alias{normalPhenotype}
\title{Simulate normally distributed phenotype}
\usage{
normalPhenotype(means, standardDeviations, phenotypeName, marker)
}
\arguments{
\item{means}{The means of the phenotype for all the different founder alleles}

\item{standardDeviations}{The standard deviations of the phenotype for all the different founder alleles}

\item{phenotypeName}{The name of the new phenotype}

\item{marker}{The name of the marker which controls this phenotype}
}
\value{
An object of class \code{normalPhenotype} representing the phenotype.
}
\description{
Add a normally distributed phenotype
}
\details{
Add a normally distributed phenotype to a given populations
}
