% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_threads.R
\name{omp_set_num_threads}
\alias{omp_set_num_threads}
\alias{omp_get_num_threads}
\title{Get or set number of threads for OpenMP}
\usage{
omp_set_num_threads(num)

omp_get_num_threads()
}
\arguments{
\item{num}{New number of threads for OpenMP}
}
\value{
None

The number of threads for OpenMP
}
\description{
Get or set the number of threads for OpenMP
}
\details{
Some functions in mpMap2 are parallelised. Depending on the number of cores available, and the type of workload, it may be advantageous to turn parallelisation on or off, by setting the number of OpenMP threads appropriately. Setting the number of threads to 1 turns parallelisation off

In particular, for small examples on a computer with a large number of threads, parallelisation may result in a huge decrease in performance. 

This function returns an error if the package was not compiled with OpenMP.
}
