% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeGraph-class.R
\docType{class}
\name{pedigreeGraph-class}
\alias{pedigreeGraph-class}
\alias{.pedigreeGraph}
\title{Graph for a pedigree}
\description{
Graph for a pedigree
}
\details{
This class contains the directed graph corresponding to a pedigree, and data for laying out the graph on a plane.
}
\section{Slots}{

\describe{
\item{\code{graph}}{An object of class igraph.}

\item{\code{layout}}{A matrix where each row gives the position of a graph vertex in the plane.}
}}

