% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeToGraph.R
\name{pedigreeToGraph}
\alias{pedigreeToGraph}
\title{Convert pedigree to a graph}
\usage{
pedigreeToGraph(pedigree)
}
\arguments{
\item{pedigree}{The pedigree to convert into a graph}
}
\value{
An object of class \code{pedigreeGraph}, containing the graph and a planar layout for the graph.
}
\description{
Convert pedigree to a graph
}
\details{
It is often useful for visualisation purposes to generate the pedigree graph. In this graph, every genetic line is a vertex in a graph, and there is an edge from every parent to all the offspring. This function generates the graph, and lays the graph out in the plane in a way that tends to make the structure of the graph as clear as possible.
}
