% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addExtraMarkers.R
\name{plot,addExtraMarkersStatistics,ANY-method}
\alias{plot,addExtraMarkersStatistics,ANY-method}
\title{Plot chi-squared statistics for independence}
\usage{
\S4method{plot}{addExtraMarkersStatistics,ANY}(x, y, ...)
}
\arguments{
\item{x}{Object of class \code{addExtraMarkersStatistics} containing test-statistic values.}

\item{y}{Unused}

\item{...}{Unused}
}
\value{
A ggplot object suitable for display.
}
\description{
Plot the chi-squared statistics for independence, used to map a new marker to an existing genetic map
}
\details{
This function plots a trace of the chi-squared test-statistics used to map a new genetic marker to an existing genetic map. This can be useful to, for example, see if a single polymorphism is present at multiple points on the genome.
}
