% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,mpcross-method}
\alias{plot,mpcross-method}
\title{Plot methods}
\usage{
\S4method{plot}{mpcross}(x, y, ...)
}
\arguments{
\item{x}{Unused}

\item{y}{Unused}

\item{...}{Unused}
}
\description{
There are multiple meaningful ways to plot mpMap2 objects. Please use \code{\link{plotProbabilities}} or \code{\link{plotMosaic}} instead.
}
\details{
There are multiple meaningful ways to plot mpMap2 objects. Please use \code{\link{plotProbabilities}} or \code{\link{plotMosaic}} instead.
}
