% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeGraph.R
\name{plot,pedigreeGraph,ANY-method}
\alias{plot,pedigreeGraph,ANY-method}
\title{Plot the graph of a pedigree}
\usage{
\S4method{plot}{pedigreeGraph,ANY}(x, y, ...)
}
\arguments{
\item{x}{pedigree graph to plot}

\item{y}{unused}

\item{...}{Other options to plot.igraph}
}
\value{
None
}
\description{
Plot the graph of a pedigree
}
\details{
Plot the graph of a pedigree, after the graph has been generated by \code{\link{pedigreeToGraph}}
}
\seealso{
\code{\link{pedigreeToGraph}}
}
