% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeGraph.R
\docType{methods}
\name{plot,pedigreeGraph-method}
\alias{plot,pedigreeGraph-method}
\title{Plot the graph of a pedigree}
\usage{
\S4method{plot}{pedigreeGraph}(x, y, ...)
}
\description{
Plot the graph of a pedigree
}
\details{
Plot the graph of a pedigree, after the graph has been generated by \code{\link{pedigreeToGraph}}
}
\seealso{
\code{\link{pedigreeToGraph}}
}
