% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,mpcross-method}
\alias{plot,mpcross-method}
\alias{plot,geneticData-method}
\alias{plot,probabilities-method}
\alias{plot,imputed-method}
\title{Plot methods}
\usage{
\S4method{plot}{mpcross}(x, y, ...)

\S4method{plot}{geneticData}(x, y, ...)

\S4method{plot}{probabilities}(x, y, ...)

\S4method{plot}{imputed}(x, y, ...)
}
\arguments{
\item{x}{Unused}

\item{y}{Unused}

\item{...}{Unused}
}
\description{
There are multiple meaningful ways to plot some mpMap2 objects. Please use \code{\link{plotProbabilities}} or \code{\link{plotMosaic}} instead.
}
\details{
There are multiple meaningful ways to plot some mpMap2 objects. In these cases the plot function is implemented but returns an error. Please use \code{\link{plotProbabilities}} or \code{\link{plotMosaic}} instead.
}
