% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMosaic.R
\name{plotMosaic}
\alias{plotMosaic}
\title{Plot estimated genetic composition of lines}
\usage{
plotMosaic(inputObject, chromosomes, positions, lines, ...)
}
\arguments{
\item{inputObject}{An object of class \code{mpcrossMapped} containing imputed IBD genotypes}

\item{chromosomes}{Chromosomes to plot}

\item{positions}{Genetic positions to plot}

\item{lines}{Genetic lines to plot}

\item{...}{Extra inputs to \code{heatmap_2}}
}
\value{
None
}
\description{
Plot estimated genetic composition of lines
}
\details{
This function produces a heatmap showing the genetic composition of lines, as measured by the imputed IBD genotypes. Rows correspond to genetic lines, columns correspond to genetic positions, and colours indicate founder alleles. All heterozygotes are marked in the same colour, otherwise there are generally too many colours to be useful.
}
