% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact.R
\docType{methods}
\name{redact}
\alias{redact}
\alias{redact,mpcross-method}
\alias{redact,mpcrossRF-method}
\alias{redact,mpcrossLG-method}
\alias{redact,mpcrossMapped-method}
\alias{redact,geneticData-method}
\title{Redact sensitive information

This function redacts possibly sensitive information from objects, resulting in an object that is safe to publish.}
\usage{
redact(object)

\S4method{redact}{mpcross}(object)

\S4method{redact}{mpcrossRF}(object)

\S4method{redact}{mpcrossLG}(object)

\S4method{redact}{mpcrossMapped}(object)

\S4method{redact}{geneticData}(object)
}
\description{
Sensitive information includes names of genetic lines (both founding lines and final population lines) and marker names. All actual data (marker genotypes, imputed IBD genotypes, IBD probabilities, etc) are preserved.
}
