% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rilPedigree.R
\name{rilPedigree}
\alias{rilPedigree}
\title{Generate a two-parent RIL pedigree which starts from inbred founders}
\usage{
rilPedigree(populationSize, selfingGenerations)
}
\arguments{
\item{populationSize}{The size of the generated population}

\item{selfingGenerations}{Number of generations of selfing. Specifying one generation leads to an F2 design.}
}
\value{
An object of class \code{detailedPedigree} representing the experimental design, suitable for simulation using simulateMPCross.
}
\description{
Generate a two-parent RIL pedigree which starts from inbred founders
}
