% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMPCross.R
\name{simulateMPCross}
\alias{simulateMPCross}
\title{Simulate data from multi-parent designs}
\usage{
simulateMPCross(map, pedigree, mapFunction, seed)
}
\arguments{
\item{map}{Linkage map with which to generate data. See \code{\link[qtl]{sim.map}}}

\item{pedigree}{Pedigree for a multi-parent cross.}

\item{mapFunction}{Map function used to convert distances to recombination fractions}

\item{seed}{Random seed to use.}
}
\value{
Object of class \code{mpcross}.
}
\description{
Data is simulated according to a pedigree, map and QTL model
}
