% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toMpMap.R
\name{toMpMap}
\alias{toMpMap}
\title{Convert to mpMap format}
\usage{
toMpMap(mpcross)
}
\arguments{
\item{mpcross}{The object of class \code{mpcross} to convert.}
}
\value{
An object with structure compatible with the older mpMap package.
}
\description{
Convert to the format used by the original mpMap package.
}
\details{
Converts an \code{mpcross} object to the format used by the original mpMap, the predecessor of this package. It is unlikely that this function will ever need to be used.
}
