% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addExtraMarkerFromRawCall.R
\name{addExtraMarkerFromRawCall}
\alias{addExtraMarkerFromRawCall}
\title{Add an extra marker from raw calling data}
\usage{
addExtraMarkerFromRawCall(mpcrossMapped, newMarker,
  useOnlyExtraImputationPoints = TRUE)
}
\arguments{
\item{mpcrossMapped}{An existing dataset with a map, which must include imputation data and recombination fraction data.}

\item{useOnlyExtraImputationPoints}{Use only the additional points at which imputation data has been generated? Currently this must be \code{TRUE}.}

\item{newData}{A data.frame object containing the raw data for the marker to add}
}
\value{
A vector containing the test statistic values.
}
\description{
Add an extra marker to a map, based on raw calling data, using a QTL-mapping style approach.
}
\details{
This function uses a QTL-mapping style approach to test for where an extra marker should be added to an existing map. The code uses the imputation data at a collection of points, and the \emph{raw calling data} for the extra marker. Test statistics are computed using a multivariate analysis of variance approach. If the imputed genotype at a point is independent of the calling data, then that marker probably should not be mapped to that point. If the imputed genotype at a point and the calling data are strongly dependent, then the marker should probably be mapped to that point. Dependence and independence are measured using an F-test. 

Currently the set of points which are tested is the set of points at which imputation data is available, \emph{which are not markers}. The intention is that this set of points should be an equally spaced grid of points; this has the affect of radically reducing the number of tests that are performed, as generally there are far fewer points in this grid, than there are markers. As the position chosen will need to be inspected and changed manually in any case, any loss in accuracy by using the grid of point is essentially irrelevant. In future it may be possible to use every marker position as the set of points at which tests are performed, by setting \code{useOnlyExtraImputationPoints} to \code{FALSE}.
}

