% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignFounderPattern.R
\name{assignFounderPattern}
\alias{assignFounderPattern}
\title{Set founder genotypes}
\usage{
assignFounderPattern(founderMatrix)
}
\arguments{
\item{founderMatrix}{The new matrix of founder genotypes}
}
\value{
An object of internal class \code{assignFounderPattern}, suitable for application to an object of class \code{mpcross} using the addition operation.
}
\description{
Set founder genotypes
}
\details{
Set the founder genotypes to a specified matrix, for an object with fully informative markers. This can allow the same set of founder genotypes to be used for multiple simulation runs.
}
