% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/founders.R
\name{founderNames}
\alias{founderNames}
\alias{founderNames,mpcross-method}
\alias{founderNames,geneticData-method}
\title{Names of founding genetic lines

Return the names of the founding genetic lines

If the \code{mpcross} object contains a single experiment a vector of names of genetic lines is returned. If an \code{mpcross} object contains multiple experiments a list of vectors of names is returned.}
\usage{
founderNames(object)

\S4method{founderNames}{mpcross}(object)

\S4method{founderNames}{geneticData}(object)
}
\arguments{
\item{object}{The \code{mpcross} object from which to extract the names of the founding genetic lines}
}
\value{
A vector of names of genetic lines, or a list of such vectors, in the case of multiple experiments.
}
\description{
Names of founding genetic lines

Return the names of the founding genetic lines

If the \code{mpcross} object contains a single experiment a vector of names of genetic lines is returned. If an \code{mpcross} object contains multiple experiments a list of vectors of names is returned.
}
