% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markers}
\alias{markers}
\alias{markers,mpcross-method}
\alias{markers,geneticData-method}
\alias{markers,rf-method}
\alias{markers,lg-method}
\alias{markers,hetData-method}
\title{Genotyped markers

Return the names of the genotyped markers. 

If an \code{mpcross} object contains multiple experiments, all experiments are required to have the same markers. So a single vector of marker names is returned, in all cases.}
\usage{
markers(object)

\S4method{markers}{mpcross}(object)

\S4method{markers}{geneticData}(object)

\S4method{markers}{rf}(object)

\S4method{markers}{lg}(object)

\S4method{markers}{hetData}(object)
}
\arguments{
\item{object}{The \code{mpcross} object from which to extract the marker names}
}
\value{
The names of the genetic markers.
}
\description{
Genotyped markers

Return the names of the genotyped markers. 

If an \code{mpcross} object contains multiple experiments, all experiments are required to have the same markers. So a single vector of marker names is returned, in all cases.
}
