% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nFounders.R
\name{nFounders}
\alias{nFounders}
\alias{nFounders,detailedPedigree-method}
\alias{nFounders,pedigree-method}
\alias{nFounders,mpcross-method}
\alias{nFounders,geneticData-method}
\title{Number of genotyped markers

Return the number of genotyped markers in an object.

If an \code{mpcross} object contains multiple experiments, one number is returned per experiment.}
\usage{
nFounders(object)

\S4method{nFounders}{detailedPedigree}(object)

\S4method{nFounders}{pedigree}(object)

\S4method{nFounders}{mpcross}(object)

\S4method{nFounders}{geneticData}(object)
}
\arguments{
\item{object}{The \code{mpcross} object from which to extract the number of founders}
}
\value{
The number of founding lines in the population, or a list of numbers in the case of multiple experiments contained in a single object.
}
\description{
Number of genotyped markers

Return the number of genotyped markers in an object.

If an \code{mpcross} object contains multiple experiments, one number is returned per experiment.
}
