% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,mpcross,ANY-method}
\alias{plot,mpcross,ANY-method}
\alias{plot,geneticData,ANY-method}
\alias{plot,probabilities,ANY-method}
\alias{plot,imputed,ANY-method}
\title{Plot methods}
\usage{
\S4method{plot}{mpcross,ANY}(x, y, ...)

\S4method{plot}{geneticData,ANY}(x, y, ...)

\S4method{plot}{probabilities,ANY}(x, y, ...)

\S4method{plot}{imputed,ANY}(x, y, ...)
}
\arguments{
\item{x}{Unused}

\item{y}{Unused}

\item{...}{Unused}
}
\value{
None
}
\description{
There are multiple meaningful ways to plot some mpMap2 objects. Please use \code{\link{plotProbabilities}} or \code{\link{plotMosaic}} instead.
}
\details{
There are multiple meaningful ways to plot some mpMap2 objects. In these cases the plot function is implemented but returns an error. Please use \code{\link{plotProbabilities}} or \code{\link{plotMosaic}} instead.
}
